+++ g2g-plugins/src/lib.rs
@@ -9,6 +9,7 @@
 
 extern crate alloc;
 
+pub mod capsfilter;
 pub mod fakesink;
 pub mod h264parse;
 pub mod identity;